Add TPC-H into CI perf testing
Added docker image with patched Tarantool for running TPC-H bench.
Added TPC-H bench into CI perf testing.

Need to avoid of use external variables as much as possible.
It is already exists IMAGE_PERF_BUILT variable and the new
one for TPCH should be based on it.

Build image with patched Tarantool for TPC-H bench set to not
fatal errors on its fails, due to the image uses Tarantool
sources patching and should be updated each time sources changed
and patching fails, so to avoid of fails of the images preparation
process this build should be blocked from failing.

Co-authored-by: Alexander V. Tikhonov <avtikhon@tarantool.org>

Part of tarantool/tarantool#5010