durability: direct tpcc_load usage, removed sed
load.sh script that was used before has been wirtten in such a manner
that it was hard to detect any errors during it's execution (not to
mention the fact that tpcc_load does not exit on any error).
Unfortunately this patch depends on a newer version of tarantool/tpcc
that exits on repetable errors, uses correct schemas for the latest
tarantool version and uses correct API from the latest tarantool
versions.